+++ helper.h
@@ -12,10 +12,10 @@ void algorithm(int k, int d, int num_rows, int max_iter, double **data_points, d
 void free_array_2d(double **arr, int r);
 void set_equal_array_2d(double **new, double **current, int r, int c);
 void zero_array_2d(double **arr, int r, int c);
-void calculate_new_centroids(int k,int d,const double *num_elements_in_cluster,double** new_centroids, double **clusters);
+void calculate_new_centroids(int k,int d,const int *num_elements_in_cluster,double** new_centroids, double **clusters);
 double **allocate_array_2d(int r, int c);
 double calculate_distance_squared(double *a, double *b, int d);
 void error_occurred(int condition);
-void set_clusters(int k,int d, int num_rows, double *num_elements_in_cluster,double **centroids, double **clusters, double **data_points);
+void set_clusters(int k,int d, int num_rows, int *num_elements_in_cluster,double **centroids, double **clusters, double **data_points);
 
 #endif
